// src/app/interfaces/req-favoritos-response-interfaces.ts
// Generated by https://quicktype.io

export interface FavoritoResponse {
    HEADER:   Header;
    RESPONSE: EventoFavorito[];
}

export interface Header {
    CODE:          number;
    CODE_RESPONSE: string;
    MESSAGE:       string;
}

export interface EventoFavorito {
    esfavorito:     false;
    favorito:     null;
    idfavoritos:       number;
    evento:            number;
    fecha:             number;
    usuario:           number;
    estado:            number;
    ideventos:         number;
    titulo:            string;
    categoria_id:      number;
    EsGratis:          string;
    direccion:         string;
    Numero:            string;
    latitud_longitud:  string;
    tipoMoneda:        string;
    descripcionEvento: string;
    Departamento:      string;
    Distrito:          string;
    Provincia:         string;
    NombreLocal:       string;
    DescripcionLocal:  string;
    Destacado:         number;
    urlFuente:         string;
    tipoEntrada:       number;
    usuario_id:        number;
    created_at:        string;
    updated_at:        null;
    idimagenes:        number;
    evento_id:         number;
    nombreImagen:      string;
    url:               string;
    formato:           string;
    Tipo:              null;
    idfecha:           number;
    FechaInicio:       string;
    Monto:             number;
}